Fix many linter errors

+++ src/RealEngine/renderer/PhysDeviceSuitability.cpp
@@ -15,6 +15,7 @@ bool areExtensionsSupported(vk::PhysicalDevice physicalDevice) {
     std::bitset<k_deviceExtensions.size()> supported{};
     for (const auto& ext : physicalDevice.enumerateDeviceExtensionProperties()) {
         for (size_t i = 0; i < k_deviceExtensions.size(); ++i) {
+            // NOLINTNEXTLINE(*-array-index): i is in range
             if (std::strcmp(ext.extensionName.data(), k_deviceExtensions[i]) == 0) {
                 supported[i] = true;
             }
@@ -89,6 +90,7 @@ OutStruct* toOutStruct(std::byte* ptr) {
     return reinterpret_cast<OutStruct*>(ptr);
 }
 
+// NOLINTBEGIN(*-c-arrays)
 std::unique_ptr<std::byte[]> copyPhysDeviceCreateInfoChain(const void* chain) {
     assert(chain);
     assert(toInStruct(chain)->sType == vk::StructureType::ePhysicalDeviceFeatures2);
@@ -115,12 +117,13 @@ std::unique_ptr<std::byte[]> copyPhysDeviceCreateInfoChain(const void* chain) {
 
     return buf;
 }
+// NOLINTEND(*-c-arrays)
 
 int findPreferred(
     std::span<const vk::PhysicalDevice> physDevices, std::string_view preferredDevice
 ) {
     if (!preferredDevice.empty()) {
-        for (size_t i = 0; i < physDevices.size(); ++i) {
+        for (int i = 0; i < static_cast<int>(physDevices.size()); ++i) {
             const auto& physDev = physDevices[i];
             auto props          = physDev.getProperties2().properties;
             std::string_view name{props.deviceName};
